Update hero logic + State Machine + new assets

+++ Assets/Scripts/GameLogic/Hero/Data/HeroFighterData.cs
@@ -7,7 +7,7 @@ namespace CastleWarriors.GameLogic.Data
     {
         [field: SerializeField] public float Damage { get; private set; }
         [field: SerializeField] public float DelayBetweenAttack { get; private set; }
-        [field: SerializeField] public float AttackCounts { get; private set; }
+        [field: SerializeField] public int AttackCounts { get; private set; }
         [field: SerializeField] public float TriggerDistance { get; private set; }
         [field: SerializeField] public float DistanceToStartAttack {get; private set;}
     }
